django-server/requirements.txt
pip3 freeze                                                                                                                                                                                    [4886189] 
asgiref==3.3.1
bson==0.5.8
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
Django==2.2.17
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangotoolbox==1.8.0
djongo==1.3.0
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.3
pycparser==2.20
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
rest-meets-djongo==0.0.13
six==1.15.0
sqlparse==0.2.4
Werkzeug==1.0.1